backend/requirements.txt
asgiref==3.2.3
attrs==19.3.0
backcall==0.1.0
certifi==2019.11.28
chardet==3.0.4
coreapi==2.3.3
coreschema==0.0.4
decorator==4.4.1
Django==3.0.2
django-cors-headers==3.2.1
django-dotenv==1.4.2
django-extensions==2.2.5
django-filter==2.2.0
djangorestframework==3.11.0
drf-yasg==1.17.0
idna==2.8
inflection==0.3.1
ipython==7.11.1
ipython-genutils==0.2.0
itypes==1.1.0
jedi==0.15.2
Jinja2==2.10.3
jsonschema==3.2.0
Markdown==3.1.1
MarkupSafe==1.1.1
packaging==20.0
parso==0.5.2
pexpect==4.7.0
pickleshare==0.7.5
prompt-toolkit==3.0.2
psycopg2-binary==2.8.4
ptyprocess==0.6.0
Pygments==2.5.2
pyparsing==2.4.6
pyrsistent==0.15.7
pytz==2019.3
PyYAML==5.3
requests==2.22.0
ruamel.yaml==0.16.5
six==1.13.0
sqlparse==0.3.0
swagger-spec-validator==2.4.3
traitlets==4.3.3
uritemplate==3.0.1
urllib3==1.25.7
wcwidth==0.1.8
